+++ flutter_quill_extensions/lib/embeds/toolbar/camera_button.dart
@@ -1,10 +1,11 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_quill/flutter_quill.dart' hide Text;
 import 'package:image_picker/image_picker.dart';
 
-import '../../models/themes/quill_icon_theme.dart';
-import '../../translations/toolbar.i18n.dart';
-import '../controller.dart';
-import '../toolbar.dart';
+import 'package:flutter_quill/translations.dart';
+
+import '../embed_types.dart';
+import 'image_video_utils.dart';
 
 class CameraButton extends StatelessWidget {
   const CameraButton({