+++ flutter_quill_extensions/lib/embeds/toolbar/image_button.dart
@@ -1,13 +1,12 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_quill/flutter_quill.dart';
+import 'package:image_picker/image_picker.dart';
 
-import '../../models/documents/nodes/embeddable.dart';
-import '../../models/themes/quill_dialog_theme.dart';
-import '../../models/themes/quill_icon_theme.dart';
-import '../controller.dart';
-import '../toolbar.dart';
+import '../embed_types.dart';
+import 'image_video_utils.dart';
 
-class FormulaButton extends StatelessWidget {
-  const FormulaButton({
+class ImageButton extends StatelessWidget {
+  const ImageButton({
     required this.icon,
     required this.controller,
     this.iconSize = kDefaultIconSize,
@@ -60,9 +59,44 @@ class FormulaButton extends StatelessWidget {
   }
 
   Future<void> _onPressedHandler(BuildContext context) async {
-    final index = controller.selection.baseOffset;
-    final length = controller.selection.extentOffset - index;
+    if (onImagePickCallback != null) {
+      final selector =
+          mediaPickSettingSelector ?? ImageVideoUtils.selectMediaPickSetting;
+      final source = await selector(context);
+      if (source != null) {
+        if (source == MediaPickSetting.Gallery) {
+          _pickImage(context);
+        } else {
+          _typeLink(context);
+        }
+      }
+    } else {
+      _typeLink(context);
+    }
+  }
 
-    controller.replaceText(index, length, BlockEmbed.formula(''), null);
+  void _pickImage(BuildContext context) => ImageVideoUtils.handleImageButtonTap(
+        context,
+        controller,
+        ImageSource.gallery,
+        onImagePickCallback!,
+        filePickImpl: filePickImpl,
+        webImagePickImpl: webImagePickImpl,
+      );
+
+  void _typeLink(BuildContext context) {
+    showDialog<String>(
+      context: context,
+      builder: (_) => LinkDialog(dialogTheme: dialogTheme),
+    ).then(_linkSubmitted);
+  }
+
+  void _linkSubmitted(String? value) {
+    if (value != null && value.isNotEmpty) {
+      final index = controller.selection.baseOffset;
+      final length = controller.selection.extentOffset - index;
+
+      controller.replaceText(index, length, BlockEmbed.image(value), null);
+    }
   }
 }