+++ flutter_quill_extensions/lib/embeds/toolbar/image_video_utils.dart
@@ -2,15 +2,13 @@ import 'dart:io';
 
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_quill/flutter_quill.dart' hide Text;
 import 'package:image_picker/image_picker.dart';
 
-import '../../models/documents/nodes/embeddable.dart';
-import '../../models/rules/insert.dart';
-import '../../models/themes/quill_dialog_theme.dart';
-import '../../translations/toolbar.i18n.dart';
-import '../../utils/platform.dart';
-import '../controller.dart';
-import '../toolbar.dart';
+import 'package:flutter_quill/translations.dart';
+import 'package:flutter_quill/extensions.dart';
+
+import '../embed_types.dart';
 
 class LinkDialog extends StatefulWidget {
   const LinkDialog({this.dialogTheme, this.link, Key? key}) : super(key: key);
@@ -73,13 +71,6 @@ class LinkDialogState extends State<LinkDialog> {
   void _applyLink() {
     Navigator.pop(context, _link.trim());
   }
-}
-
-enum MediaPickSetting {
-  Gallery,
-  Link,
-  Camera,
-  Video,
 }
 
 class ImageVideoUtils {