+++ flutter_quill_extensions/lib/embeds/toolbar/video_button.dart
@@ -1,11 +1,9 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_quill/flutter_quill.dart';
 import 'package:image_picker/image_picker.dart';
 
-import '../../models/documents/nodes/embeddable.dart';
-import '../../models/themes/quill_dialog_theme.dart';
-import '../../models/themes/quill_icon_theme.dart';
-import '../controller.dart';
-import '../toolbar.dart';
+import '../embed_types.dart';
+import 'image_video_utils.dart';
 
 class VideoButton extends StatelessWidget {
   const VideoButton({