+++ flutter_quill_extensions/lib/embeds/widgets/youtube_video_app.dart
@@ -1,9 +1,8 @@
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_quill/flutter_quill.dart';
 import 'package:url_launcher/url_launcher.dart';
 import 'package:youtube_player_flutter_quill/youtube_player_flutter_quill.dart';
-
-import '../default_styles.dart';
 
 class YoutubeVideoApp extends StatefulWidget {
   const YoutubeVideoApp(